refactor: unify guest page layout
