+++ src/pages/GuestTimeline.tsx
@@ -1,14 +1,10 @@
 import { useEffect, useRef, useState } from 'react'
-import { Box, Button, CssBaseline, Divider, Paper, ThemeProvider, darken } from '@mui/material'
-import type { ConcurrentTheme } from '../model'
-import { Link, useLocation, useParams } from 'react-router-dom'
+import { Box, Divider, Paper } from '@mui/material'
+import { useLocation, useParams } from 'react-router-dom'
 import { Timeline } from '../components/Timeline/main'
 import { Client, type User } from '@concurrent-world/client'
 import { FullScreenLoading } from '../components/ui/FullScreenLoading'
 import ApiProvider from '../context/api'
-import { Themes, loadConcurrentTheme } from '../themes'
-import { usePersistent } from '../hooks/usePersistent'
-import { ConcurrentWordmark } from '../components/theming/ConcurrentWordmark'
 import TickerProvider from '../context/Ticker'
 
 import { type VListHandle } from 'virtua'
@@ -18,6 +14,7 @@ import ListIcon from '@mui/icons-material/List'
 import AlternateEmailIcon from '@mui/icons-material/AlternateEmail'
 import { Profile } from '../components/Profile'
 import { MessageContainer } from '../components/Message/MessageContainer'
+import { GuestBase } from '../components/GuestBase'
 
 export interface GuestPageProps {
     page: 'stream' | 'entity' | 'message'
@@ -82,158 +79,88 @@ export function GuestTimelinePage(props: GuestPageProps): JSX.Element {
         }
     }, [props.page, id, reactlocation.hash])
 
-    const [themeName, setThemeName] = usePersistent<string>('Theme', 'sacher')
-    const [theme, setTheme] = useState<ConcurrentTheme>(loadConcurrentTheme(themeName))
-    const themes: string[] = Object.keys(Themes)
-    const randomTheme = (): void => {
-        const box = themes.filter((e) => e !== themeName)
-        const newThemeName = box[Math.floor(Math.random() * box.length)]
-        setThemeName(newThemeName)
-        setTheme(loadConcurrentTheme(newThemeName))
-    }
-
     const scrollParentRef = useRef<HTMLDivElement>(null)
 
     if (!client) return <FullScreenLoading message="Loading..." />
 
     return (
-        <ThemeProvider theme={theme}>
-            <CssBaseline />
+        <GuestBase
+            sx={{
+                display: 'flex',
+                flexDirection: 'column',
+                flexGrow: 1,
+                overflow: 'hidden',
+                padding: 1
+            }}
+        >
             <TickerProvider>
                 <ApiProvider client={client}>
-                    <Box
-                        sx={{
-                            display: 'flex',
-                            justifyContent: 'center',
-                            background: [
-                                theme.palette.background.default,
-                                `linear-gradient(${theme.palette.background.default}, ${darken(
-                                    theme.palette.background.default,
-                                    0.1
-                                )})`
-                            ],
-                            width: '100vw',
-                            height: '100dvh'
-                        }}
-                    >
-                        <Box
+                    <>
+                        {props.page === 'message' && (
+                            <Paper
+                                sx={{
+                                    margin: { xs: 0.5, sm: 1 },
+                                    display: 'flex',
+                                    flexFlow: 'column',
+                                    p: 2
+                                }}
+                            >
+                                <MessageContainer
+                                    messageID={id?.split('@')[0] ?? ''}
+                                    messageOwner={id?.split('@')[1] ?? ''}
+                                />
+                            </Paper>
+                        )}
+
+                        <Paper
                             sx={{
+                                flexGrow: '1',
+                                margin: { xs: 0.5, sm: 1 },
+                                mb: { xs: 0, sm: '10px' },
                                 display: 'flex',
-                                flex: 1,
-                                maxWidth: '1280px',
-                                width: '100%'
+                                flexFlow: 'column',
+                                borderRadius: 2,
+                                overflow: 'hidden',
+                                background: 'none'
                             }}
                         >
                             <Box
                                 sx={{
+                                    width: '100%',
+                                    minHeight: '100%',
+                                    backgroundColor: 'background.paper',
                                     display: 'flex',
-                                    flexFlow: 'column',
-                                    overflow: 'hidden',
-                                    flex: 1
+                                    flexDirection: 'column'
                                 }}
                             >
-                                <Box display="flex" justifyContent="space-between" mt={2} mx={2}>
-                                    <Button
-                                        disableRipple
-                                        variant="text"
-                                        sx={{
-                                            display: 'flex',
-                                            justifyContent: 'center',
-                                            alignItems: 'center',
-                                            flexDirection: 'row',
-                                            textTransform: 'none',
-                                            '&:hover': {
-                                                background: 'none'
-                                            }
-                                        }}
-                                        component={Link}
-                                        to="/welcome"
-                                    >
-                                        <ConcurrentWordmark color={theme.palette.background.contrastText} />
-                                    </Button>
-                                    <Box
-                                        sx={{
-                                            display: 'flex',
-                                            gap: '10px'
-                                        }}
-                                    >
-                                        <Button onClick={randomTheme}>✨</Button>
-                                        <Button component={Link} to="/register">
-                                            はじめる
-                                        </Button>
-                                    </Box>
-                                </Box>
+                                <TimelineHeader title={title} titleIcon={id ? <AlternateEmailIcon /> : <ListIcon />} />
 
-                                {props.page === 'message' && (
-                                    <Paper
-                                        sx={{
-                                            margin: { xs: 0.5, sm: 1 },
-                                            display: 'flex',
-                                            flexFlow: 'column',
-                                            p: 2
-                                        }}
-                                    >
-                                        <MessageContainer
-                                            messageID={id?.split('@')[0] ?? ''}
-                                            messageOwner={id?.split('@')[1] ?? ''}
-                                        />
-                                    </Paper>
-                                )}
-
-                                <Paper
-                                    sx={{
-                                        flexGrow: '1',
-                                        margin: { xs: 0.5, sm: 1 },
-                                        mb: { xs: 0, sm: '10px' },
-                                        display: 'flex',
-                                        flexFlow: 'column',
-                                        borderRadius: 2,
-                                        overflow: 'hidden',
-                                        background: 'none'
-                                    }}
-                                >
-                                    <Box
-                                        sx={{
-                                            width: '100%',
-                                            minHeight: '100%',
-                                            backgroundColor: 'background.paper',
-                                            display: 'flex',
-                                            flexDirection: 'column'
-                                        }}
-                                    >
-                                        <TimelineHeader
-                                            title={title}
-                                            titleIcon={id ? <AlternateEmailIcon /> : <ListIcon />}
-                                        />
-
-                                        <Timeline
-                                            ref={timelineRef}
-                                            streams={targetStream}
-                                            header={
-                                                <Box
-                                                    sx={{
-                                                        overflowX: 'hidden',
-                                                        overflowY: 'auto',
-                                                        overscrollBehaviorY: 'contain'
-                                                    }}
-                                                    ref={scrollParentRef}
-                                                >
-                                                    {user && (
-                                                        <>
-                                                            <Profile user={user} id={id} guest={true} />
-                                                            <Divider />
-                                                        </>
-                                                    )}
-                                                </Box>
-                                            }
-                                        />
-                                    </Box>
-                                </Paper>
+                                <Timeline
+                                    ref={timelineRef}
+                                    streams={targetStream}
+                                    header={
+                                        <Box
+                                            sx={{
+                                                overflowX: 'hidden',
+                                                overflowY: 'auto',
+                                                overscrollBehaviorY: 'contain'
+                                            }}
+                                            ref={scrollParentRef}
+                                        >
+                                            {user && (
+                                                <>
+                                                    <Profile user={user} id={id} guest={true} />
+                                                    <Divider />
+                                                </>
+                                            )}
+                                        </Box>
+                                    }
+                                />
                             </Box>
-                        </Box>
-                    </Box>
+                        </Paper>
+                    </>
                 </ApiProvider>
             </TickerProvider>
-        </ThemeProvider>
+        </GuestBase>
     )
 }
